requirements.txt
paddlepaddle==0.10.0
